move shared tool helpers to tool-helpers.ts and out of the tools/ folder
