+++ src/tool-helpers.ts
