+++ src/tools/tasks-list-completed.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod'
 import type { TodoistTool } from '../todoist-tool'
-import { mapTask } from './shared'
+import { mapTask } from '../tool-helpers'
 
 const ArgsSchema = {
     getBy: z
