+++ src/tools/tasks-search.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod'
 import type { TodoistTool } from '../todoist-tool'
-import { getTasksByFilter } from './shared'
+import { getTasksByFilter } from '../tool-helpers'
 
 const ArgsSchema = {
     searchText: z.string().min(1).describe('The text to search for in tasks.'),
